Clean up index.js / create app structure
broke up the clutter that was the original index.js by created three main files: tools, bot, and input.

In order to do this, also created a micro back end with node express to host the static html and expose the public folder to the client.
