+++ README.md
@@ -2,6 +2,10 @@
 
 ApostleAI is a conversation bot modeled after Paul the Apostle, commonly known as Saint Paul, an apostle who spread the teachings of Jesus in the first-century world.
 
+to run the bot install `NodeJS` then enter project directory and hit it with a `npm install` which should install all dependencies, in this case the only one necessary should be [express](https://github.com/expressjs/express)
+
+Once the dependencies have been installed `npm run start` will run the server and host the application to `:8080`
+
 [View the live website here!](https://prophet-ai.github.io/ApostleAI/)
 
 ![alt text](https://upload.wikimedia.org/wikipedia/commons/thumb/6/67/The_Predication_of_Saint_Paul_LACMA_M.2000.179.24.jpg/800px-The_Predication_of_Saint_Paul_LACMA_M.2000.179.24.jpg)
