+++ index.html
@@ -42,7 +42,7 @@
 	</div>
 
   <script src="js/scramble.js"></script>
-  <script src="js/index.js"></script>
+  <script src="js/index.js" type="module"></script>
   
   <script>
     const phrases = [
