+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "apostleai",
+  "version": "1.0.0",
+  "description": "A wise albeit mysterious bot and client to ease your weary mind",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/prophet-ai/ApostleAI.git"
+  },
+  "author": "apostle-ai-team",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/prophet-ai/ApostleAI/issues"
+  },
+  "homepage": "https://github.com/prophet-ai/ApostleAI#readme",
+  "dependencies": {
+    "express": "^4.17.1"
+  }
+}
